// src/components/customs/dropdown/dropdown-types.ts
type TurboDropdownProperties = TurboProperties & {
    values: (string | DropdownEntry)[];
    selectedValues?: string[],

    selector?: string | HTMLElement;
    popup?: HTMLElement;

    multiSelection?: boolean;
    underlyingInputName?: string,

    customSelectorTag?: keyof HTMLElementTagNameMap;
    customEntryTag?: keyof HTMLElementTagNameMap;

    customSelectorClasses?: string;
    customPopupClasses?: string;
    customEntriesClasses?: string;
    customSelectedEntriesClasses?: string;
};
// ...
type TurboDropdownConfig = {
    defaultEntryTag?: keyof HTMLElementTagNameMap;
    defaultSelectorTag?: keyof HTMLElementTagNameMap;

    defaultSelectorClasses?: string | string[]
    defaultPopupClasses?: string | string[];
    defaultEntriesClasses?: string | string[];
    defaultSelectedEntriesClasses?: string | string[];
}